refactor: extract slack-events types + reaction-router from index.ts (#76)
The v0.13-v0.14 work already split the SlackAdapter into ~15 modules; two more
clean pieces come out here, behaviour-preserving:

- slack-events.ts — the ambient Socket-Mode payload shapes, converted from a
  local `declare namespace Slack` into an exported `Slack` namespace so the
  adapter and its handlers share one source (index.ts imports the type).
- reaction-router.ts — `routeReactionAdded`, the emoji→action dispatch
  (:cr: review / :a: approve / 📚 atom-save / ✅❌ approve-reject / 🎫 issue /
  👎 citation-questioned). Coordinators + web client + config passed in as deps;
  index.ts's handleReactionAdded is now a thin delegate. Removes the knowledge/
  telemetry/events imports it alone used.

index.ts 1314 → 1127 lines. Guarded end-to-end: slack-adapter.test.ts drives
reaction_added via socket.emit; 1009 cli tests pass, typecheck clean.

The message/app_mention/dm/slash-command handlers stay on the adapter — they're
deeper in the live path and warrant host live-verification before extraction.
