+++ apps/docs/docs/changelog.md
@@ -30,6 +30,7 @@ Merged to `main`, not yet tagged.
   - `review.ts` 1186 → 1028 lines. The coordinator keeps the claim lifecycle, progress bar, and Slack replies; the extracted seam still warrants host live-verification of the end-to-end `:cr:`/`:a:` path before the next release.
 - **admin.ts decomposition** ([#76](https://github.com/hanfour/pm-workspace-kit/issues/76)) — the `/pmk admin` dispatcher pulled its two largest command clusters into domain modules: `admin-review.ts` (`review …` provider/strategy/approval config) and `admin-audience.ts` (`audience …` per-user/channel/default tiers + v0.15 examples), over a shared `admin-shared.ts` (request/result types, `logAdmin`, the Slack-mention parsers) that breaks the dispatcher↔handler cycle. `admin.ts` 827 → 445 lines; `handleAdminSlash` behaviour unchanged (guarded by its existing 57 test call-sites, `extractUserId`/`extractChannelId`/`RuntimeHealthSnapshot` re-exported).
 - **mra.ts decomposition** ([#76](https://github.com/hanfour/pm-workspace-kit/issues/76)) — pulled the review-protocol contract and the secret-strip out of the 1163-line adapter: `mra-env.ts` (`strippedChildEnv` + the secret-shaped-env denylist, shared by every mra spawn) and `mra-review-protocol.ts` (`MraReviewResult` / strategy+provider types, `buildReviewArgv`, `parseReviewStdout`, `reviewEnv` — pure, no spawning). `runMraReview` keeps the spawn wiring; both modules re-exported from `mra.ts` for back-compat. `mra.ts` 1163 → 1038 lines; behaviour-preserving (1009 cli tests, the argv/parse/env contract already had heavy coverage in mra-review.test.ts).
+- **slack/index.ts decomposition** ([#76](https://github.com/hanfour/pm-workspace-kit/issues/76)) — beyond the v0.13–v0.14 extractions, pulled two more pieces out of the `SlackAdapter`: `slack-events.ts` (the ambient Socket-Mode payload types, now an exported `Slack` namespace) and `reaction-router.ts` (`routeReactionAdded` — the emoji→action dispatch for `:cr:`/`:a:`/📚/✅/❌/🎫/👎, taking the coordinators + web client as deps). `index.ts` 1314 → 1127 lines; behaviour-preserving, guarded end-to-end by slack-adapter.test.ts's `reaction_added` socket-emit tests. The remaining event handlers (message / app_mention / dm / slash-command) stay on the adapter — they warrant host live-verification before further extraction.
 
 ### Security
